Move `ObviousFloat` to a separate module.
